mainly made cosmetic changes, moved some functions into they own files, changes optparse to argparse, since optparse is deprecated. Also changes how contacts details are handled. There is always going to be default contact email address, however user can pass they own email address and name, no defaul name is set.
