+++ templates/check.html
@@ -1,0 +1,41 @@
+<!doctype html>
+<html>
+<head>
+<title>nothhing</title>
+</head>
+
+<body>
+		<h1 class="text-center">{{title}}</h1>
+	
+		{% if not no_contact %}
+		<!-- PUT YOUR AUTHOR HERE -->
+		<p class="text-center">ContaCt: {{author_name}} (<a href="mailto:{{author_email}}">{{author_email}}</a>)</p>
+                {% endif %}
+
+		{% if not no_contents %}		
+		<!-- AUTOGENERATED TABLE OF CONTENTS.. NO NEED TO TOUCH -->
+		<h2>Contents</h2>
+		<div class="panel">
+			<div id="toc">
+			
+			</div>
+		</div>  
+                {% endif %}
+		
+		<!-- REPORT SUMMARY HERE -->
+		{% if summary %}
+		<h2>Summary</h2>
+      	<div class="panel">
+	        <p>
+                <strong>{{summary}}</strong>
+        </div>		
+		{% endif %}
+
+
+               {{ content }}
+
+		
+</body>
+
+</html>
+
